feat(cdk/testing): fix circular dep
