+++ src/cdk/testing/public-api.ts
@@ -11,3 +11,4 @@ export * from './harness-environment';
 export * from './test-element';
 export * from './element-dimensions';
 export * from './text-filtering';
+export * from './change-detection';
